chore: Cleanup components, part of #1318
