+++ src/components/common/contentful-content.module.scss
@@ -2,7 +2,7 @@
   color: white;
   background: $color-plum-700;
 }
-.edit-raw {
+.edit-inline {
   @extend %edit-color;
   padding: spacer(4) spacer(8);
 }
