--- src/components/common/list-arrow.module.scss
@@ -1,24 +1,0 @@
-.list-arrow[class] {
-  @include type-size(400);
-  line-height: 1.5;
-  margin: spacer(24) 0;
-}
-.list-arrow-item {
-  flex-wrap: nowrap;
-  display: flex;
-  flex-direction: row-reverse;
-  margin-bottom: spacer(8);
-
-  .list-arrow-icon {
-    flex: 0 0 toRem(30);
-    color: $color-plum-700;
-    @include type-size(400);
-  }
-  p {
-    flex: 1;
-    margin: 0;
-  }
-  a {
-    font-weight: bold;
-  }
-}
