--- src/components/common/section-divider.module.scss
@@ -1,10 +1,0 @@
-.divider {
-  height: 1px;
-  background-color: $color-slate-200;
-  border: none;
-  @include margin(64, (top, bottom));
-
-  @media (max-width: $viewport-md) {
-    margin: 3rem 0; // ignore-style-rule
-  }
-}
