+++ src/pages/about/index.js
@@ -4,7 +4,7 @@ import Container from '~components/common/container'
 import ContentfulContent from '~components/common/contentful-content'
 import LongContent from '~components/common/long-content'
 import Layout from '~components/layout'
-import VolunteersList from '~components/common/volunteers-list'
+import VolunteersList from '~components/pages/about/volunteers-list'
 
 export default ({ data }) => (
   <Layout title="About Us" path="/about">
